requirements.txt
Django==1.10.3
django-bower==5.2.0
django-bootstrap3
whitenoise==2.0.6
Pillow
django-crispy-forms==1.6.0
python-magic==0.4.12
psycopg2==2.6.2
django-contact-form==1.3
